requirements.txt
# Application dependencies
pypdf
python-docx
nltk